requirements.txt
# Python 3.10+ recommended
pandas>=2.0
python-dotenv>=1.0
# Standard library only for email sending (smtplib, ssl, email.message)
# No additional pip packages required for SMTP
